Refactor of mount API (mostly renaming and moving classes)
